+++ pukr/__init__.py
@@ -1,4 +1,4 @@
 from .config import get_logger
 
-__version__ = '0.0.1'
-__update_date__ = '2022-04-24_15:41:22'
+__version__ = '0.0.2'
+__update_date__ = '2022-05-01_19:59:20'